refactor: Move main url keeping function out of the main
